+++ test-e2e.sh
@@ -46,6 +46,7 @@ source "$E2E_DIR/test_board_daemon.sh"
 source "$E2E_DIR/test_task_comments.sh"
 source "$E2E_DIR/test_task_copy.sh"
 source "$E2E_DIR/test_task_archive.sh"
+source "$E2E_DIR/test_hook_attention.sh"
 
 ################################################################################
 # Main Test Execution
@@ -75,6 +76,7 @@ main() {
     test_task_comments
     test_task_copy
     test_task_archive
+    test_hook_attention
 
     # Cleanup
     cleanup_test_db